+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Features and improvements
 
+- Add option `viewport-glyph` to `text-rotation-alignment` which places glyphs along a linestring and rotates them to the x-axis of the viewport (#716).
+
 - *...Add new stuff here...*
 
 ### 🐞 Bug fixes
@@ -9,6 +11,8 @@
 - Change `GeoJSONFeature.id` type from `number | string | void` to `number | string | undefined` (#1093)
 - Add FeatureIdentifier type to define feature parameter in setFeatureState, removeFeatureState, and getFeatureState methods. Change FeatureIdentifier.id from `id: string | number;` to `id?: string | number | undefined;` (#1095)
 - Change map.on, map.off, and map.once type parameter from "type: MapEvent" to "type: MapEvent | string" (#1094)
+
+- *...Add new stuff here...*
 
 ## 2.1.7